+++ src/ewatercycle/plugins/hype/forcing.py
@@ -17,10 +17,18 @@ class HypeForcing(DefaultForcing):
     """Container for hype forcing data.
 
     Args:
-        Pobs (str): Input file for precipitation data.
-        TMAXobs (str): Input file for maximum temperature data.
-        TMINobs (str): Input file for minimum temperature data.
-        Tobs (str): Input file for temperature data.
+        directory: Directory where forcing data files are stored.
+        start_time: Start time of forcing in UTC and ISO format string e.g.
+            'YYYY-MM-DDTHH:MM:SSZ'.
+        end_time: End time of forcing in UTC and ISO format string e.g.
+            'YYYY-MM-DDTHH:MM:SSZ'.
+        shape: Path to a shape file. Used for spatial selection.
+        Pobs (optional): Input file for precipitation data. Defaults to 'Pobs.txt'
+        TMAXobs (optional): Input file for maximum temperature data. Defaults to
+            'TMAXobs.txt'
+        TMINobs (optional): Input file for minimum temperature data. Defaults to
+            'TMINobs.txt'
+        Tobs (optional): Input file for temperature data. Defaults to 'Tobs.txt'
     """
 
     # type ignored because pydantic wants literal in base class while mypy does not
@@ -91,7 +99,7 @@ class HypeForcing(DefaultForcing):
             ds.to_csv(p, sep=" ", index_label="DATE", float_format="%.3f")
 
         # instantiate forcing object based on generated data
-        return HypeForcing(
+        generated_forcing = HypeForcing(
             directory=directory,
             start_time=start_time,
             end_time=end_time,
@@ -101,3 +109,5 @@ class HypeForcing(DefaultForcing):
             TMINobs=forcing_files["TMINobs"].name,
             Tobs=forcing_files["Tobs"].name,
         )
+        generated_forcing.save()
+        return generated_forcing